pipeline/pyproject.toml
[project]
name = "certipass-pipeline"
version = "0.1.0"
description = "AI pipeline to generate quiz JSON from exam text"
requires-python = ">=3.11"
dependencies = [
    "langchain>=0.3",
    "langchain-openai>=0.3",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "ruff",
]
